refactor(tui): move checklist history renderer
Extract checklist/todo snapshot parsing and compact transcript rendering from history.rs into history/checklist.rs while preserving the existing test-facing helper surface.

Refs #3308.
